migrate: move resourcesSupport to products directory
- Migrated resourcesSupport product files to src/products/resourcesSupport/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved query.ts from src/data/queries/ to src/products/resourcesSupport/
- Moved test file and renamed resourcesSupport.test.tsx -> query.test.ts
- Moved formatted type from src/types/formatted/ to src/products/resourcesSupport/formatted-type.ts
- Moved mock file from src/mocks/ to src/products/resourcesSupport/mock.js
- Updated all import statements across the codebase
- Fixed import paths in migrated files (formatted-type.ts, query.ts, template.test.tsx, query.test.ts)
- All TypeScript errors resolved
- All tests passing (3 query tests + 1 template test)
- Updated migration guide with completed status
- Added CODEOWNERS entry for new product directory
- Verified dev server runs and resourcesSupport pages load correctly
