+++ .github/CODEOWNERS
@@ -28,3 +28,4 @@ src/products/vamcSystem/ @department-of-veterans-affairs/ap-team @department-of-
 src/products/healthCareLocalFacility/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
 src/products/vamcHealthServicesListing/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
 src/products/locationsListing/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
+src/products/resourcesSupport/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
