+++ src/data/queries/index.ts
@@ -39,7 +39,7 @@ import * as QaParagraph from '@/data/queries/qaParagraph'
 import * as QaSection from '@/data/queries/qaSection'
 import * as QuestionAnswer from '@/data/queries/questionAnswer'
 import * as ReactWidget from '@/data/queries/reactWidget'
-import * as ResourcesSupport from '@/data/queries/resourcesSupport'
+import * as ResourcesSupport from '@/products/resourcesSupport/query'
 import * as StaticPathResources from '@/data/queries/staticPathResources'
 import * as StoryListing from '@/products/storyListing/query'
 import * as SupportServices from '@/data/queries/supportServices'
