+++ src/pages/[[...slug]].tsx
@@ -48,7 +48,7 @@ import { LocationsListing as FormattedLocationsListing } from '@/products/locati
 import { NewsStory as FormattedNewsStory } from '@/products/newsStory/formatted-type'
 import { PressRelease as FormattedPressRelease } from '@/products/pressRelease/formatted-type'
 import { PressReleaseListing as FormattedPressReleaseListing } from '@/products/pressReleaseListing/formatted-type'
-import { ResourcesSupport as FormattedResourcesSupport } from '@/types/formatted/resourcesSupport'
+import { ResourcesSupport as FormattedResourcesSupport } from '@/products/resourcesSupport/formatted-type'
 import { StaffProfile as FormattedStaffProfile } from '@/products/staffProfile/formatted-type'
 import { StoryListing as FormattedStoryListing } from '@/products/storyListing/formatted-type'
 import { VetCenter as FormattedVetCenter } from '@/products/vetCenter/formatted-type'
@@ -70,7 +70,7 @@ import { NewsStory } from '@/products/newsStory/template'
 import { PressRelease } from '@/products/pressRelease/template'
 import { PressReleaseListing } from '@/products/pressReleaseListing/template'
 import { PreviewCrumb } from '@/templates/common/preview'
-import { ResourcesSupport } from '@/templates/layouts/resourcesSupport'
+import { ResourcesSupport } from '@/products/resourcesSupport/template'
 import { StaffProfile } from '@/products/staffProfile/template'
 import { StoryListing } from '@/products/storyListing/template'
 import { VetCenter } from '@/products/vetCenter/template'
